emotion_model/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
opencv-python>=4.8.0
numpy==1.26.4
matplotlib>=3.5.0
seaborn>=0.12.0
scikit-learn>=1.0.0

